+++ graph/graph_test.go
@@ -6,36 +6,36 @@ func testGraph(t *testing.T, graph Graph) {
 	testCounts(t, graph, 0, 0)
 
 	for i := 1; i <= 10; i++ {
-		graph.AddNode(int64(i), interface{}(int64(i)))
+		graph.AddNode(i, interface{}(i))
 	}
 
 	testCounts(t, graph, 10, 0)
 
 	for i := 1; i <= 10; i++ {
-		_, d := graph.Node(int64(i))
-		if d != int64(i) {
+		_, d := graph.Node(i)
+		if d != i {
 			t.Errorf("Expected node %d to hold %d, was %d", i, i, d)
 		}
 	}
 
 	for i := 1; i <= 10; i++ {
-		edges := graph.Edges(int64(i))
+		edges := graph.Edges(i)
 		if len(edges) > 0 {
 			t.Errorf("Expected empty edges initially for %d, was %v", i, edges)
 		}
 	}
 
-	graph.RemoveNode(int64(10))
-	graph.RemoveNode(int64(5))
+	graph.RemoveNode(10)
+	graph.RemoveNode(5)
 
 	testCounts(t, graph, 8, 0)
 
-	graph.AddEdge(int64(1), int64(2), int64(5))
-	graph.AddEdge(int64(2), int64(3), int64(10))
+	graph.AddEdge(1, 2, 5)
+	graph.AddEdge(2, 3, 10)
 
 	testCounts(t, graph, 8, 2)
 
-	edges := graph.Edges(int64(2))
+	edges := graph.Edges(2)
 	if len(edges) != 1 {
 		t.Errorf("Expected node 2 to have 1 edge, had %d edges", len(edges))
 	}
@@ -47,15 +47,15 @@ func testGraph(t *testing.T, graph Graph) {
 	if !ok {
 		t.Errorf("Expected graph to have edge from node 1 to node 2 prior to removal")
 	}
-	if w != int64(5) {
+	if w != 5 {
 		t.Errorf("Expected edge from node 1 to node 2 to have weight 5, was %d", w)
 	}
 
-	graph.RemoveEdge(int64(1), int64(2))
+	graph.RemoveEdge(1, 2)
 
 	testCounts(t, graph, 8, 1)
 
-	edges = graph.Edges(int64(1))
+	edges = graph.Edges(1)
 	if len(edges) != 0 {
 		t.Errorf("Expected node 1 to have no edges after its edge was removed, had %v", edges)
 	}
